+++ android/src/main/kotlin/com/umair/beacons_plugin/BeaconHelper.kt
@@ -9,7 +9,9 @@ import io.flutter.plugin.common.EventChannel
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import org.altbeacon.beacon.*
+import timber.log.Timber
 import java.util.*
+import kotlin.math.roundToInt
 
 
 open class BeaconHelper(var context: Context) : BeaconConsumer, BeaconsPlugin.Companion.PluginImpl {
@@ -32,6 +34,21 @@ open class BeaconHelper(var context: Context) : BeaconConsumer, BeaconsPlugin.Co
 
         @JvmStatic
         private var listOfRegions = arrayListOf<Region>()
+
+        @JvmStatic
+        private var listOfBeaconLayouts = arrayListOf<String>()
+
+        @JvmStatic
+        private var foregroundScanPeriod = 1100L
+
+        @JvmStatic
+        var foregroundBetweenScanPeriod = 0L
+
+        @JvmStatic
+        var backgroundScanPeriod = 1100L
+
+        @JvmStatic
+        var backgroundBetweenScanPeriod = 0L
     }
 
     private val TAG = "BeaconHelper"
@@ -43,8 +60,28 @@ open class BeaconHelper(var context: Context) : BeaconConsumer, BeaconsPlugin.Co
         isBluetoothEnabled(context)
     }
 
+    override fun addBeaconLayout(layout: String) {
+        listOfBeaconLayouts.add(layout)
+    }
+
+    override fun setBackgroundScanPeriod(
+        backgroundScanPeriod: Long,
+        backgroundBetweenScanPeriod: Long
+    ) {
+        BeaconHelper.backgroundScanPeriod = backgroundScanPeriod
+        BeaconHelper.backgroundBetweenScanPeriod = backgroundBetweenScanPeriod
+    }
+
+    override fun setForegroundScanPeriod(
+        foregroundScanPeriod: Long,
+        foregroundBetweenScanPeriod: Long
+    ) {
+        BeaconHelper.foregroundScanPeriod = foregroundScanPeriod
+        BeaconHelper.foregroundBetweenScanPeriod = foregroundBetweenScanPeriod
+    }
+
     override fun stopMonitoringBeacons() {
-        Log.i(TAG, "stopMonitoringBeacons")
+        Timber.i(TAG, "stopMonitoringBeacons")
 
         beaconManager?.unbind(this)
         eventSink = null
@@ -54,16 +91,15 @@ open class BeaconHelper(var context: Context) : BeaconConsumer, BeaconsPlugin.Co
         this.eventSink = events
     }
 
-
     override fun onBeaconServiceConnect() {
-        Log.i(TAG, "onBeaconServiceConnect")
+        Timber.i(TAG, "onBeaconServiceConnect")
 
         beaconManager?.removeAllMonitorNotifiers()
 
         beaconManager?.addMonitorNotifier(object : MonitorNotifier {
             override fun didEnterRegion(region: Region) {
                 try {
-                    Log.d(TAG, "didEnterRegion")
+                    Timber.d(TAG, "didEnterRegion")
                     beaconManager?.startRangingBeaconsInRegion(region)
                 } catch (e: RemoteException) {
                     e.printStackTrace()
@@ -73,7 +109,7 @@ open class BeaconHelper(var context: Context) : BeaconConsumer, BeaconsPlugin.Co
 
             override fun didExitRegion(region: Region) {
                 try {
-                    Log.d(TAG, "didExitRegion")
+                    Timber.d(TAG, "didExitRegion")
                     beaconManager?.stopRangingBeaconsInRegion(region)
                 } catch (e: RemoteException) {
                     e.printStackTrace()
@@ -125,19 +161,19 @@ open class BeaconHelper(var context: Context) : BeaconConsumer, BeaconsPlugin.Co
 
         //Distance
         val distance1 = b.distance
-        val distance = (Math.round(distance1 * 100.0) / 100.0).toString()
+        val distance = ((distance1 * 100.0).roundToInt() / 100.0).toString()
 
         val message = Beacon(
-                name = identifier,
-                uuid = uuid,
-                major = major,
-                minor = minor,
-                macAddress = address,
-                distance = distance,
-                rssi = rssi,
-                txPower = txPower,
-                proximity = Beacon.getProximityOfBeacon(b).value,
-                scanTime = getReadableTime(System.currentTimeMillis())
+            name = identifier,
+            uuid = uuid,
+            major = major,
+            minor = minor,
+            macAddress = address,
+            distance = distance,
+            rssi = rssi,
+            txPower = txPower,
+            proximity = Beacon.getProximityOfBeacon(b).value,
+            scanTime = getReadableTime(System.currentTimeMillis())
         ).toString()
 
         //Log.i(TAG, "sendBeaconData: $message")
@@ -147,27 +183,39 @@ open class BeaconHelper(var context: Context) : BeaconConsumer, BeaconsPlugin.Co
 
     private fun startMonitoringBeacons(region: Region) {
         try {
-            Log.i(TAG, "startMonitoringBeacons: ${region.uniqueId}")
+            Timber.i(TAG, "startMonitoringBeacons: ${region.uniqueId}")
             beaconManager?.startMonitoringBeaconsInRegion(region)
             beaconManager?.startRangingBeaconsInRegion(region)
         } catch (e: RemoteException) {
             e.printStackTrace()
-            Log.e(TAG, e.message.toString())
+            Timber.e(TAG, e.message.toString())
         }
     }
 
     private fun setUpBeaconManager(context: Context) {
         if (BeaconsPlugin.permissionsGranted(context)) {
 
-            Log.i(TAG, "setUpBeaconManager")
+            Timber.i(TAG, "setUpBeaconManager")
             beaconManager = BeaconManager.getInstanceForApplication(context)
+
             beaconManager?.beaconParsers?.add(BeaconParser().setBeaconLayout("m:2-3=beac,i:4-19,i:20-21,i:22-23,p:24-24,d:25-25"))
             beaconManager?.beaconParsers?.add(BeaconParser().setBeaconLayout("m:2-3=0215,i:4-19,i:20-21,i:22-23,p:24-24"))
-            beaconManager?.backgroundBetweenScanPeriod = 0
-            beaconManager?.backgroundScanPeriod = 1100
+
+            listOfBeaconLayouts.forEach { layout ->
+                if (layout.isNotEmpty()) {
+                    beaconManager?.beaconParsers?.add(BeaconParser().setBeaconLayout(layout))
+                }
+            }
+
+            beaconManager?.backgroundBetweenScanPeriod = backgroundBetweenScanPeriod
+            beaconManager?.backgroundScanPeriod = backgroundScanPeriod
+
+            beaconManager?.foregroundBetweenScanPeriod = foregroundBetweenScanPeriod
+            beaconManager?.foregroundScanPeriod = foregroundScanPeriod
+
             beaconManager?.bind(this)
         } else {
-            Log.e(TAG, "Location permissions are needed.")
+            Timber.e(TAG, "Location permissions are needed.")
         }
     }
 
@@ -188,14 +236,14 @@ open class BeaconHelper(var context: Context) : BeaconConsumer, BeaconsPlugin.Co
         listOfRegions.add(region)
 
         result.success("Region Added: ${region.uniqueId}, UUID: ${region.id1}")
-        Log.i(TAG, "Region Added: ${region.uniqueId}, UUID: ${region.id1}")
+        Timber.i(TAG, "Region Added: ${region.uniqueId}, UUID: ${region.id1}")
     }
 
     override fun clearRegions(call: MethodCall, result: MethodChannel.Result) {
         listOfRegions = arrayListOf<Region>()
 
         result.success("Regions Cleared")
-        Log.i(TAG, "Regions Cleared")
+        Timber.i(TAG, "Regions Cleared")
     }
 
     override fun startScanning() {
@@ -204,12 +252,12 @@ open class BeaconHelper(var context: Context) : BeaconConsumer, BeaconsPlugin.Co
         setUpBeaconManager(context)
 
         if (listOfRegions.isNotEmpty()) {
-            Log.i(TAG, "Started Monitoring ${listOfRegions.size} regions.")
+            Timber.i(TAG, "Started Monitoring ${listOfRegions.size} regions.")
             listOfRegions.forEach {
                 startMonitoringBeacons(it)
             }
         } else {
-            Log.i(TAG, "startScanning: No regions added..")
+            Timber.i(TAG, "startScanning: No regions added..")
         }
     }
 }